+++ Assets/Scripts/ScriptableObjects/Comment.cs
@@ -20,7 +20,7 @@ public class Comment : SuperScriptableObject
     public int Views { get; private set; }
 
     [field: SerializeField]
-    public int Damage { get; private set; }
+    public float Damage { get; private set; }
 
     [field: SerializeField]
     public float Deley { get; private set; }
@@ -67,7 +67,7 @@ public class Comment : SuperScriptableObject
         if(int.TryParse(param["Views"], out var aux0))
             Views = aux0;
 
-        if (int.TryParse(param["Damage"], out var aux3))
+        if (float.TryParse(param["Damage"], out var aux3))
             Damage = aux3;
 
         if (float.TryParse(param["Deley"], out var aux2))
